Actually run Vim, and scrape text
Actually run Vim, capture its output over a PTY,
and scrape the text we need.

Establish a side channel for passing information
from Vim to Syncat [via the xterm window title],
and use said channel to find out the useful number
of lines in Vim's output, so we can keep only them,
and discard the rest.

This enables us to only keep the data we need, while
still passing all input directly to Vim for processing.

Set up Vim's stdin and stderr so they point to
Syncat's stdin/stderr, to match the expected behavior
when running pipelines on the command line.

Signed-off-by: Vangelis Koukis <vkoukis@gmail.com>